requirements.txt
# Core
httpx == 0.28.1
pydantic == 2.11.3
python-dotenv == 1.1.0

# DEV
pytest == 8.3.5
pytest-asyncio == 0.26.0
pytest-cov == 6.1.1